+++ Sources/BraintreeCore/BTHTTP.swift
@@ -18,7 +18,6 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
     /// DispatchQueue on which asynchronous code will be executed. Defaults to `DispatchQueue.main`.
     var dispatchQueue: DispatchQueue = DispatchQueue.main
     let baseURL: URL
-    let cacheDateValidator: BTCacheDateValidator = BTCacheDateValidator()
     var clientAuthorization: ClientAuthorization?
 
     /// Session exposed for testing
@@ -28,7 +27,6 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
         
         let delegateQueue: OperationQueue = OperationQueue()
         delegateQueue.name = "com.braintreepayments.BTHTTP"
-        delegateQueue.maxConcurrentOperationCount = OperationQueue.defaultMaxConcurrentOperationCount
         
         return URLSession(configuration: configuration, delegate: self, delegateQueue: delegateQueue)
     }()
@@ -99,15 +97,11 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
 
     // MARK: - HTTP Methods
 
-    func get(_ path: String, parameters: Encodable? = nil, shouldCache: Bool = false, completion: @escaping RequestCompletion) {
+    func get(_ path: String, parameters: Encodable? = nil, completion: @escaping RequestCompletion) {
         do {
             let dict = try parameters?.toDictionary()
             
-            if shouldCache {
-                httpRequestWithCaching(method: "GET", path: path, parameters: dict, completion: completion)
-            } else {
-                httpRequest(method: "GET", path: path, parameters: dict, completion: completion)
-            }
+            httpRequest(method: "GET", path: path, parameters: dict, completion: completion)
         } catch let error {
             completion(nil, nil, error)
         }
@@ -137,44 +131,6 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
 
     // MARK: - HTTP Method Helpers
 
-    func httpRequestWithCaching(
-        method: String,
-        path: String,
-        parameters: [String: Any]? = [:],
-        completion: RequestCompletion?
-    ) {
-        createRequest(method: method, path: path, parameters: parameters) { request, error in
-            guard let request = request else {
-                self.handleRequestCompletion(data: nil, request: nil, shouldCache: false, response: nil, error: error, completion: completion)
-                return
-            }
-
-            var cachedResponse: CachedURLResponse? = URLCache.shared.cachedResponse(for: request) ?? nil
-
-            if self.cacheDateValidator.isCacheInvalid(cachedResponse ?? nil) {
-                URLCache.shared.removeAllCachedResponses()
-                cachedResponse = nil
-            }
-
-            // The increase in speed of API calls with cached configuration caused an increase in "network connection lost" errors.
-            // Adding this delay allows us to throttle the network requests slightly to reduce load on the servers and decrease connection lost errors.
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                if let cachedResponse = cachedResponse {
-                    self.handleRequestCompletion(data: cachedResponse.data, request: nil, shouldCache: false, response: cachedResponse.response, error: nil, completion: completion)
-                } else {
-                    self.session.dataTask(with: request) { [weak self] data, response, error in
-                        guard let self else {
-                            completion?(nil, nil, BTHTTPError.deallocated("BTHTTP"))
-                            return
-                        }
-
-                        handleRequestCompletion(data: data, request: request, shouldCache: true, response: response, error: error, completion: completion)
-                    }.resume()
-                }
-            }
-        }
-    }
-
     func httpRequest(
         method: String,
         path: String,
@@ -183,17 +139,17 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
     ) {
         createRequest(method: method, path: path, parameters: parameters) { request, error in
             guard let request = request else {
-                self.handleRequestCompletion(data: nil, request: nil, shouldCache: false, response: nil, error: error, completion: completion)
+                self.handleRequestCompletion(data: nil, request: nil, response: nil, error: error, completion: completion)
                 return
             }
-
+            
             self.session.dataTask(with: request) { [weak self] data, response, error in
                 guard let self else {
                     completion?(nil, nil, BTHTTPError.deallocated("BTHTTP"))
                     return
                 }
 
-                handleRequestCompletion(data: data, request: request, shouldCache: false, response: response, error: error, completion: completion)
+                handleRequestCompletion(data: data, request: request, response: response, error: error, completion: completion)
             }.resume()
         }
     }
@@ -316,7 +272,6 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
     func handleRequestCompletion(
         data: Data?,
         request: URLRequest?,
-        shouldCache: Bool,
         response: URLResponse?,
         error: Error?,
         completion: RequestCompletion?
@@ -363,15 +318,6 @@ class BTHTTP: NSObject, NSCopying, URLSessionDelegate {
                 callCompletionAsync(with: completion, body: nil, response: nil, error: error)
             }
             return
-        }
-
-        // We should only cache the response if we do not have an error and status code is 2xx
-        let successStatusCode: Bool = httpResponse.statusCode >= 200 && httpResponse.statusCode < 300
-
-        if request != nil && shouldCache && successStatusCode, let request = request {
-            let cachedURLResponse: CachedURLResponse = CachedURLResponse(response: response, data: data)
-
-            URLCache.shared.storeCachedResponse(cachedURLResponse, for: request)
         }
 
         callCompletionAsync(with: completion, body: json, response: httpResponse, error: nil)